Reorganise components into hierarchy
