+++ src/JsonValidator/MultiscaleArrays/ZarrArray/Cube3D.svelte
