+++ src/JsonValidator/Well/ImageContainer.svelte
@@ -1,5 +1,5 @@
 <script>
-  import { getJson, validate } from "./utils";
+  import { getJson, validate } from "../../utils";
 
   export let source;
   export let path;
